* cpuid.h (can_set_flag): Implement for x86_64. 	* fhandler.h (class fhandler_dev_raw): Define devbuf members taking 	numerical values as DWORD, rather than size_t. 	* fhandler_dsp.cc: Fix debug printf calls to be type safe. 	(waveOut_callback): Align parameter types to latest MSDN description. 	(waveIn_callback): Ditto. 	(fhandler_dev_dsp::Audio_out::start): Change casts in call to 	waveOutOpen according to MSDN. 	(fhandler_dev_dsp::Audio_out::parsewav): Fix a cast. 	(fhandler_dev_dsp::Audio_in::start): Change casts in call to 	waveInOpen according to MSDN. 	* fhandler_floppy.cc: Fix debug printf calls to be type safe. 	(fhandler_dev_floppy::ioctl): Add casts to be on the safe side. 	* fhandler_proc.cc: Fix debug printf calls to be type safe.

+++ winsup/cygwin/fhandler.h
@@ -806,10 +806,10 @@ class fhandler_dev_raw: public fhandler_base
  protected:
   char *devbufalloc;
   char *devbuf;
-  size_t devbufalign;
-  size_t devbufsiz;
-  size_t devbufstart;
-  size_t devbufend;
+  DWORD devbufalign;
+  DWORD devbufsiz;
+  DWORD devbufstart;
+  DWORD devbufend;
   struct status_flags
   {
     unsigned lastblk_to_read : 1;
